refactor: css 분리 및 2d home 폴더 구조 조정
